Code refactoring. Deleting unnecessary classes.
