+++ lib/game_two_players.rb
@@ -1,4 +1,4 @@
-class GameRPSSLOnePlayer
+class GameTwoPlayers
 
   RULES = { rock: [:scissors, :lizard],
           paper: [:rock, :spock],
@@ -6,24 +6,16 @@ class GameRPSSLOnePlayer
           lizard: [:paper, :spock],
           spock: [:rock, :scissors] }
 
-  attr_reader :thrown
-
-  def initialize(choice)
-    @choice =  choice ? choice.to_sym : :rock
+  def initialize(choice_one, choice_two)
+    @choice_one = choice_one ? choice_one.to_sym : :rock
+    @choice_two = choice_two ? choice_two.to_sym : :rock
     @options = [:rock, :paper, :scissors, :spock, :lizard]
   end
 
   def won?
-    @thrown = throw
-    return :tie if @choice == @thrown
-    return :won if RULES[@choice].include? @thrown
+    return :tie if @choice_one == @choice_two
+    return :won if RULES[@choice_one].include? @choice_two
     :lose
-  end
-
-  private
-
-  def throw
-    @options.sample
   end
 
 end
